#1848 moved remediation endpoint implementation to separate file
